add physical_simulation
